Birthday's Managing program
Using pandas, Calendar, and csv modules and libraries, I build a events managing problem. First, it will ask the month and year and then displays calendar; and then it will ask you to enter the event along with the date and saving it in a csv file format. You access it whenever you want!